+++ ejs/itemsystemPage.ejs
@@ -1,7 +1,2 @@
 <ul>
-    <% data.forEach(function(item) { %>
-        <li>
-            <p><%=item%></p>
-        </li>
-    <% }) %>
 </ul>
