+++ lib/Commands/Action.js
@@ -1,8 +1,8 @@
 /**
  * Created by Stijn on 02/07/2017.
  */
-const logger = require('../model/logger'),
-    validator = require('./validator');
+const logger = require('../../model/logger'),
+    validator = require('../Helpers/validator');
 
 /**
  * Constructor for an action.
