+++ lib/Helpers/CSV-Parser.js
@@ -2,7 +2,7 @@
  * Created by Arjan on 28-6-2017.
  */
 const fs = require("fs"),
-    logger = require('../model/logger');
+    logger = require('../../model/logger');
 var blocks = [];
 var items = [];
