+++ lib/Helpers/validator.js
@@ -8,7 +8,7 @@ var validateCrafting;
 var validateAction;
 var validate;
 
-var itemMap = require("../model/itemmap");
+var itemMap = require("../../model/itemmap");
 /**
  * Calls all validate methods to validate the given commands.
  *
