+++ lib/Turtle.js
@@ -2,7 +2,7 @@
  * Created by Stijn on 02/07/2017.
  */
 const logger = require('../model/logger'),
-    validator = require('./validator');
+    validator = require('./Helpers/validator');
 /**
  * The turtle class.
  *
